// src/utils/enums.ts
export enum TransactionType {
  DEPOSIT_PIX = 'DEPOSIT_PIX',
  WITHDRAW_PIX = 'WITHDRAW_PIX',
  CONVERT_BRL_TO_USDC = 'CONVERT_BRL_TO_USDC',
  LIQUIDATE_ONCHAIN = 'LIQUIDATE_ONCHAIN'
}

export enum TransactionStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED'
}